Fixed folder structure for Front-end Pages and Components
