+++ NutriTrack-Frontend/src/Components/Sections/HomepageSections/Homepage-mid-section1.tsx
