+++ NutriTrack-Frontend/src/Pages/ContactUs.tsx
@@ -1,0 +1,9 @@
+const ContactUs = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default ContactUs
