+++ NutriTrack-Frontend/src/Pages/TermsAndConditions.tsx
@@ -1,0 +1,9 @@
+const TermsAndConditions = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default TermsAndConditions
