+++ NutriTrack-Frontend/src/Pages/index.ts
@@ -3,24 +3,25 @@ import Features from "./FeaturesPage";
 import Pricing from "./PricingPage";
 import FAQ from "./FAQPage";
 import Testimonials from "./TestimonialsPage";
-import HistoricalViewPage from './HistoricalViewPage';
+import HistoricalViewPage from './CustomerPages/HistoricalViewPage';
 import Login from './Login';
-import SignUp from './SignUp';
-import CreateCustomFoodPage from './CreateCustomFoodPage';
-import MealsConsumedPage from './MealConsumedPage';
-import TrackPage from "./TrackPage";
-import TrackCustomFoodPage from "./TrackCustomFoodPage";
-import DailyDashboardPage from "./DailyDashboardPage";
-import Dashboard from "./Dashboard";
-import ProfileSetup from "./ProfileSetup";
-import AdminDashboard from "./AdminDashboard";
+import CreateCustomFoodPage from './CustomerPages/CreateCustomFoodPage';
+import MealsConsumedPage from './CustomerPages/MealConsumedPage';
+import TrackPage from "./CustomerPages/TrackPage";
+import TrackCustomFoodPage from "./CustomerPages/TrackCustomFoodPage";
+import DailyDashboardPage from "./CustomerPages/DailyDashboardPage";
+import Dashboard from "./CustomerPages/Dashboard";
+import ProfileSetup from "./CustomerPages/ProfileSetup";
+import AdminDashboard from "./AdminPages/AdminDashboard";
 import NotFound from "./NotFound";
+import TermsAndConditions from "./TermsAndConditions";
+import AboutUs from "./AboutUs";
+import ContactUs from "./ContactUs";
 
 export {
     HomePage,
     HistoricalViewPage,
     Login,
-    SignUp,
     Features,
     Pricing,
     FAQ,
@@ -34,4 +35,7 @@ export {
     ProfileSetup,
     AdminDashboard,
     NotFound,
+    TermsAndConditions,
+    AboutUs,
+    ContactUs,
 };
